hand_detection/requirements.txt
opencv-python
numpy
pyyaml
mediapipe

